Move download code to hyoga.core.download.
